弾性体クラス Elastic を仮実装． 現在は1次元のみ対応． ヘアーシミュレーションが最も近い． 安定性向上のために，ダンパの追加が必要．

+++ trunk/Crystal/CrystalRenderer/ColorBuffer.h
@@ -23,9 +23,9 @@ public:
 		}
 	}
 
-	int getWidth() const { return colors.size(); }
+	size_t getWidth() const { return colors.size(); }
 
-	int getHeight() const { return colors.empty() ? 0 : colors[0].size(); }
+	size_t getHeight() const { return colors.empty() ? 0 : colors[0].size(); }
 
 	void clear(const Color4d& color)
 	{
